[*] Rename files from NRI to NRD
